Adds working code to get all seasons for the user
Code was added to get the seasons for the logged in user.

Add a handlebars file to display a table of available seasons
Refactores player handlebars to use the improved format used for seasons
Modified methods so that the data automatically loads when a user goes
to the create season tool.

Removed View my Seasons button as the data is now loaded automatically