Adds fixes to delete for players & seasons
Code fixes two different delete problems
seasons: modifies events.js to listen for the event.target to pass to
a the delete function rather than an the whole event.

players: fixes the div id where the event handler listens for the event

handlebars: shortens the delete button to simply: Delete to get space
back in the table buy shortening the button.