Adds code to enable auto fill and other
I admit this commit should have been broken into smaller ones.

Adds code to autofill update player form

Adds handlebars to create a selector for viewing a batting order

Adds basic form (not wired) so select a season and player

Adds code to update lists.  However, this has met with mixed success as
the server may not return a refreshed list.  Added a wait, but even then
the wait may not be long enough for the server